# app/swagger-docs.json
{"openapi":"3.0.0","paths":{"/v1/auth/login":{"post":{"description":"api key required","operationId":"AuthController_login_v1","parameters":[],"requestBody":{"required":true,"description":"Login Data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInInputDto"}}}},"responses":{"201":{"description":"Login successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInOutputDto"}}}},"401":{"description":"Login unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BusinessException"}}}}},"security":[{"api-key":[]}],"summary":"store authentication","tags":["Auth"]}},"/v1/stores":{"post":{"operationId":"StoresController_create_v1","parameters":[],"requestBody":{"required":true,"description":"Store data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateStoreInputDto"}}}},"responses":{"201":{"description":"Store created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StoreIdDto"}}}},"400":{"description":"Store has not been created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BusinessException"}}}},"409":{"description":"Store has not been created, already exists","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BusinessException"}}}}},"security":[{"api-key":[]}],"summary":"Register your store","tags":["Store"]},"get":{"description":"Retrieves the store based on the storeId contained in the accessToken (JWT).","operationId":"StoresController_findById_v1","parameters":[],"responses":{"200":{"description":"Store found successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseStoreDto"}}}},"400":{"description":"Store not found successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BusinessException"}}}}},"security":[{"access-token":[]}],"summary":"Find Store","tags":["Store"]}},"/v1/stores/totems":{"post":{"operationId":"StoresController_createTotem_v1","parameters":[],"requestBody":{"required":true,"description":"Totem data","content":{"application/json":{"schema":{"type":"object","properties":{"totemName":{"type":"string","example":"Totem 1"}},"required":["totemName"]}}}},"responses":{"201":{"description":"Totem created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseIdUuidDto"}}}},"400":{"description":"Totem has not been created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BusinessException"}}}},"409":{"description":"Totem has not been created, already exists","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BusinessException"}}}}},"security":[{"access-token":[]}],"summary":"Register the totem","tags":["Store"]}},"/v1/stores/totems/{totemId}":{"delete":{"operationId":"StoresController_deleteTotem_v1","parameters":[{"name":"totemId","required":true,"in":"path","description":"ID totem","schema":{"example":"44226bf5-0187-4d7b-b649-e460fda43b01","type":"string"}}],"responses":{"200":{"description":"Totem deleted successfully"},"404":{"description":"Totem has not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BusinessException"}}}}},"security":[{"access-token":[]}],"summary":"Delete the totem","tags":["Store"]}}},"info":{"title":"API Tech Challenge - Phase Two","description":"This application provides data from the cafeteria to customers and employees.","version":"1.0.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"access-token":{"scheme":"bearer","bearerFormat":"JWT","type":"http","name":"Authorization","in":"header"},"api-key":{"type":"apiKey","in":"header","name":"x-api-key"},"totem-token":{"type":"apiKey","in":"header","name":"totem-access-token"}},"schemas":{"SignInInputDto":{"type":"object","properties":{"email":{"type":"string","description":"E-mail field company","example":"email@gmail.com"},"password":{"type":"string","description":"Password for access","example":"Senh@1"}},"required":["email","password"]},"SignInOutputDto":{"type":"object","properties":{"access_token":{"type":"string","description":"Access Token","example":"Senh@yJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdG9yZUlkIjoiZWE2MTcxZTktOWFmNy00N2.."}},"required":["access_token"]},"BusinessException":{"type":"object","properties":{"message":{"type":"string","description":"Error description"},"statusCode":{"type":"number","description":"Code error"},"error":{"type":"string","description":"Type Error"}},"required":["message","statusCode","error"]},"CreateStoreInputDto":{"type":"object","properties":{"cnpj":{"type":"string","description":"CNPJ field company","example":"00000000000000"},"email":{"type":"string","description":"E-mail field company","example":"email@gmail.com"},"password":{"type":"string","description":"Password for access","example":"Senh@1"},"name":{"type":"string","description":"Name company","example":"Snack Bar Fiap LTDA"},"fantasy_name":{"type":"string","description":"Name company","example":"Snack Bar Fiap"},"phone":{"type":"string","description":"Contact telephone number","example":"5521999999999"}},"required":["cnpj","email","password","name","fantasy_name","phone"]},"StoreIdDto":{"type":"object","properties":{"id":{"type":"string","description":"Unique identifier for the store"}},"required":["id"]},"ResponseIdUuidDto":{"type":"object","properties":{"id":{"type":"string","description":"Unique identifier"}},"required":["id"]},"ResponseTotemDto":{"type":"object","properties":{"id":{"type":"string","description":"Unique identifier"},"name":{"type":"string","description":"Totem name","example":"Totem 01"},"tokenAccess":{"type":"string","description":"Totem access token","example":"84079b90-b1ba-4223-812f-d1f1435ea34d"}},"required":["id","name","tokenAccess"]},"ResponseStoreDto":{"type":"object","properties":{"id":{"type":"string","description":"Unique identifier for the store"},"name":{"type":"string","description":"Name company","example":"Snack Bar Fiap LTDA"},"fantasyName":{"type":"string","description":"Name company","example":"Snack Bar Fiap"},"email":{"type":"string","description":"E-mail field company","example":"email@gmail.com"},"phone":{"type":"string","description":"Contact telephone number","example":"5521999999999"},"cnpj":{"type":"string","description":"CNPJ field company","example":"00000000000000"},"totems":{"description":"Totems linked in Store","example":{"id":"84079b90-b1ba-4223-812f-d1f1435ea34d","name":"totem 01","tokenAccess":"525520790-b1ba-4223-812f-d1f1435ea34d"},"allOf":[{"$ref":"#/components/schemas/ResponseTotemDto"}]}},"required":["id","name","fantasyName","email","cnpj","totems"]}}}}